requirements.txt
flask
pandas
python-dotenv
openai
langchain
langchain-openai
langchain-experimental
pandasai
flask-cors
psycopg2-binary